Split anthropic-api.ts into separate agents
Summary:
Added the @openai/agents package for defining agentic workflows.
Converted the LLM calls in anthropic-api.ts to use this package, and
moved them into 3 separate files in the src/agents directory. This
allows us to remove some custom code for e.g. tool call result handling,
and restructures the code to be a bit cleaner in general.

Also, removed the toolUsed column from the slackMessage table because it
doesn't play well with the return types of the agents, and we'll soon
add an llm logging framework which will surface this info.

Test Plan:
- 'pnpm run eval' to make sure evals still run
- Walked through a couple flack conversations as a sanity check