Fix 'Unknown User' bug in organizeMessagesByChannelAndThread
Summary:
In some cases the bot sends messages to users which are not in the
topic.userIds array, which causes 'Unknown User' to show up in the
organizeMessagesByChannelAndThread history. Fixed this by pulling the
userIds from the channels, which eliminates the need to pass userIds as
an argument into the function.

Test Plan:
Tested manually in flack