Add sessionId and custom trace names for langfuse
Summary:
Set the groupId of openai agent traces for scheduleNextStep to
topic-<topic_id>, which becomes the sessionId for langfuse and allows us
to group traces from the same topic together. Also, set a trace name for
each trace corresponding to the name of the first agent called, so we
can see them in the list of traces.

Test Plan:
Tested with flack and saw the traces show up in langfuse UI