Fix expired refresh token, and improve findFreeSlots output (#9)
Summary:
When a user's refresh token expires (which happens after 7 days for a
non-authorized google, but 6 months of inactivity otherwise), we now
clear the user's google tokens to indicate that they need to
reauthenticate.

Also, updated the behavior of findFreeSlots so that when a user has no
calendar (either connected or manually specified), the tool will mention
that in its output, so the bot knows it doesn't have actual information
about that user's availability.

Test Plan:
Gave my test user an expired refresh token and did some testing in flack