Split out topicState table from topic to track changes over time (#29)
Summary:
We'd like to track when a topic's summary, user ids, etc change so that
we can actually replay a conversation step by step, and replicate the
exact behavior of the bot at any given point in time (for now, we're not
worried that the user contexts can also change over time -- the topic
changes more often and is more critical to track).

This allows point-in-time testing for topic routing and conversation
agents, which is becoming increasingly important as our app grows in
scope and usage and we are running into more edge cases.

To do this, we moved many of the fields from Topic to a new TopicState
table, which we now append to rather than updating. This requires a
migration, as well as substantial changes to the slack-message-handler
logic. There is also a widespread change to the codebase to use the
TopicWithState type rather than the Topic type everywhere.

Test Plan:
Ran the migration and server locally, and made sure everything went as
expected and is still working