Fix google redirect uri
Summary:
The google calendar connection redirect uri was hardcoded to
localhost:3001, which was failing for 'pnpm run dev'. It's now been
fixed to take the baseURL directly from auth.ts.

I also fixed the vite config so that it doesn't try to access the local
cert files when 'vite build' is being run, which prevents it from
erroring out on the prod server.

Test Plan:
Tested with 'pnpm run dev' locally, and was able to connect my calendar
and access the 'Calendar connected successfully!' screen