Update dump topic to return all relevant topic states
Summary:
We now return an array of states in the TopicData from the dumpTopic
function in utils.ts. This allows us to reference the correct state for
each message in the topic, rather than only the latest state in the
topic at the time that it is dumped.

If lastMessageId is specified, we also filter the states to only the
ones that have been created as of that last message, to avoid irrelevant
states that are more recent than the latest message.

As part of this, also update the Topic page on the frontend to show the
correct summary for the current message in the timeline.

Test Plan:
Tested the Topic view locally, and then dumped the test topic with two
different message ids, and saw that the correct set of states were
dumped each time