Allow adding linked github account and app installation
Summary:
Add abillity to link the github account on the profile page. We want to
link the user's github organization as well, to have access to its
repos, so to do this we had to create a custom better auth plugin that
redirects the user to the app install page instead of the default
account linking flow.

Test Plan:
Linked and unlinked the github account in local test