Add organizations, tied to slack teams (#80)
Summary:
When users log in via slack (currently the only login method), it now
assigns them to a better-auth organization based on the slack team id
they've authorized for their account. The organization is created if it
does not exist, and users cannot create organizations any other way.

Includes the addition of the better-auth organization plugin, and
corresponding migrations for the required tables.

Test Plan:
Logged in and out locally, and saw organization and membership created
in the DB