Sign out if user has no organization (#96)
Summary:
For users with existing sessions on prod, their profile page currently
throws an error, since they aren't part of any organization. We should
sign them out so that they are forced to log back in and become part of
an organization.

Test Plan:
Tested locally, deleting the organization in the database and reloading
the profile page