Add slack app installations to enable real-world usage (#100)
Summary:
Added a better auth plugin that runs through the oauth flow to add the
slack bot to your organization. Now, instead of hard-coding the slack
bot token (which is workspace-specific) to an env var, we store it in
the database as a result of this auth flow. This enables other orgs to
install pivotal to their slack workspace by going to our website and
clicking through the authentication flows.

Test Plan:
Tested locally by installing and uninstalling Pivotal Dev, and testing
that 'pnpm run dev' still works