Introduces queryAgent (#95)
Introduces queryAgent.

Summary:
The queryAgent introdues a third workflow -- queries. We used to only have two workflows -- scheduling and meeting-prep. These are quite tightly scoped, and queryAgent is a bit wider. This agent broadly introduces 4 types of capabilities:
1) Calendar connection status (is my calendar connected? is parker's calendar connected?),
2) Availability queries (what's blocking me after 3pm today? when is parker free tomorrow?),
3) Meeting lookups (send me the link again for the sync with Kai. when does my meeting when Ben start and end?),
4) Status tracking and nudges (who are we still waiting? how long has Parker been quiet? can you bump everyone who hasn't replied yet?).

In general, the queryAgent is hoping to be a catch-all for all the questions to pivotal that aren't explicitly for scheduling or meeting-prep. It will hopefully become the surface for wide-context queries, pending state tracking with GitHub integration and more work on transcripts. There is no dynamic topic switching in this commit, 'query' is not set as a default for the agent.

Test Plan:
I tested each of these functionalities extensively in dev.