Check team id when connecting google calendar via slack button
Summary:
When the user connects their calendar via the slack buttons, we append a
team parameter to the request with the current slack team id. If the
user then logs into the pivotal webapp with a different workspace, we
show them an error message on the profile page. The calendar still
connects to their user in pivotal, but in order to use it with the

This includes a migration to add a teamId column to accounts, and we now
add slack's teamId to the slack account during login.

Test Plan:
pnpm run dev