Avoid resetting slack bot token in local or dev environment (#122)
Summary:
When we click the "Uninstall Bot" button on the token page, we don't
want to invalidate the dev bot token, because it means everyone else's
dev environment will be broken until they reinstall it.

As part of this, migrated all existing environment checks to use
standard isLocalEnv, isDevEnv, and isProdEnv functions imported from
src/utils.ts.

Test Plan:
Uninstalled and reinstalled the slack bot locally, then replaced the bot
token in the DB with the old token from before the uninstall, and made
sure I could still communicated with the Pivotal Dev bot on slack