Revert to sonnet 4 to fix overzealous updateUserCalendar behavior (#132)
Summary:
Sonnet 4.5 was updating user calendars based on information in the
previous topic messages, even when explicitly instructed not to. This
led to a lot of additional LLM calls in the oneliners tests. Added the
explicit instructions not to do this, and reverted to sonnet 4, to fix
this behavior

Test Plan:
pnpm run oneliners -f benchmark_2simusers_1-25start_1-75end_60min_gen20251010131946545_eval20251010152028907_group0_topic.json -n 3