Add prompt caching for anthropic agent calls (#136)
Summary:
Added the "fetch" configuration to the OpenAI model we use for the
OpenAI agents sdk, which shims a custom fetch implementation in place
of Javascript's fetch method. In this shim, we (1) split the system
prompt into two pieces if the PROMPT_CACHE_BOUNDARY string is found in
the system prompt, and (2) add the cache_control: { type: 'ephemeral' }
param to one or both system messages, as well as the last two user or
tool messages found in the message array. For non-anthropic models, we
don't set the cache_control parameter, and just remove the
PROMPT_CACHE_BOUNDARY string rather than splitting the prompt in two.

Test Plan:
Tested locally. Logged to see that the system prompt is getting split
into two and that the cache_control param is getting set properly in the
custom fetch function. Also saw that cache_writes and cache_reads are
now affecting the cost in openrouter activity logs.